Fix all points on one vector
Of note, one of the changes in this commit modifies what the
`m_lastWrittenX` and `Y` variables get set to when first creating
a vector. They used to be set to the *tip* of the vector, but it
seems like they should be set to the *base* of the vector.